+++ Engine/PatchBank.cpp
@@ -1,6 +1,6 @@
 /*
  * mTroll MIDI Controller
- * Copyright (C) 2007-2008,2010-2016,2018,2021 Sean Echevarria
+ * Copyright (C) 2007-2008,2010-2016,2018,2021,2024 Sean Echevarria
  *
  * This file is part of mTroll.
  *
@@ -35,6 +35,7 @@
 
 
 static std::list<PatchPtr>	sActiveVolatilePatches;
+static PatchPtr sActiveProgramChangePatches[16];
 #ifdef ITEM_COUNTING
 std::atomic<int> gPatchBankCnt;
 #endif
@@ -53,6 +54,8 @@ PatchBank::PatchBank(int number,
 PatchBank::~PatchBank()
 {
 	sActiveVolatilePatches.clear();
+	for (auto &patch : sActiveProgramChangePatches)
+		patch = nullptr;
 #ifdef ITEM_COUNTING
 	--gPatchBankCnt;
 #endif
@@ -254,7 +257,10 @@ PatchBank::Load(IMainDisplay * mainDisplay, ISwitchDisplay * switchDisplay)
 					curItem->mPatch->OverridePedals(true); // expression pedals only apply to first patch
 
 				if (stA == curItem->mPatchStateAtBankLoad)
+				{
 					curItem->mPatch->BankTransitionActivation();
+					CheckForDeviceProgramChange(curItem.get(), switchDisplay, mainDisplay);
+				}
 				else if (stB == curItem->mPatchStateAtBankLoad)
 					curItem->mPatch->BankTransitionDeactivation();
 
@@ -298,7 +304,10 @@ PatchBank::Unload(IMainDisplay * mainDisplay, ISwitchDisplay * switchDisplay)
 					curItem->mPatch->OverridePedals(true);
 
 				if (stA == curItem->mPatchStateAtBankUnload)
+				{
 					curItem->mPatch->BankTransitionActivation();
+					CheckForDeviceProgramChange(curItem.get(), switchDisplay, mainDisplay);
+				}
 				else if (stB == curItem->mPatchStateAtBankUnload)
 					curItem->mPatch->BankTransitionDeactivation();
 
@@ -338,7 +347,7 @@ PatchBank::PatchSwitchPressed(SwitchFunctionAssignment st,
 {
 	PatchVect & curPatches = mPatches[switchNumber].GetPatchVect(st);
 	PatchVect::iterator it;
-	bool curSwitchHasNormalPatch = false;
+	bool curSwitchHasVolatilePatch = false;
 
 	// if any patch for the switch is normal, then need to do normal processing
 	// on current normal patches.
@@ -352,12 +361,12 @@ PatchBank::PatchSwitchPressed(SwitchFunctionAssignment st,
 
 		if (curSwitchItem->mPatch->IsPatchVolatile())
 		{
-			curSwitchHasNormalPatch = true;
+			curSwitchHasVolatilePatch = true;
 			break;
 		}
 	}
 
-	if (curSwitchHasNormalPatch)
+	if (curSwitchHasVolatilePatch)
 	{
 		// do B processing
 		for (std::list<PatchPtr>::iterator it2 = sActiveVolatilePatches.begin();
@@ -496,6 +505,8 @@ PatchBank::PatchSwitchPressed(SwitchFunctionAssignment st,
 			_ASSERTE(std::find(sActiveVolatilePatches.begin(), sActiveVolatilePatches.end(), curSwitchItem->mPatch) == sActiveVolatilePatches.end());
 			sActiveVolatilePatches.push_back(curSwitchItem->mPatch);
 		}
+
+		CheckForDeviceProgramChange(curSwitchItem.get(), switchDisplay, mainDisplay);
 
 		if (curSwitchItem->mPatch->UpdateMainDisplayOnPress())
 		{
@@ -884,6 +895,30 @@ PatchBank::ResetExclusiveGroup(ISwitchDisplay * switchDisplay,
 				continue;
 
 			curSwitchItem->mPatch->UpdateState(switchDisplay, enabled);
+		}
+	}
+}
+
+void
+PatchBank::CheckForDeviceProgramChange(BankPatchState *curSwitchItem,
+									   ISwitchDisplay *switchDisplay, 
+									   IMainDisplay *mainDisplay)
+{
+	// support for device program change tracking (actually limited to MIDI channel 
+	// rather than device) [#22]
+	const int ch = curSwitchItem->mPatch->GetDeviceProgramChangeChannel();
+	if (-1 < ch)
+	{
+		_ASSERTE(0 <= ch && 16 > ch);
+		PatchPtr oldChannelPatchItem = sActiveProgramChangePatches[ch];
+		if (oldChannelPatchItem != curSwitchItem->mPatch)
+		{
+			if (oldChannelPatchItem && oldChannelPatchItem->IsActive())
+			{
+				oldChannelPatchItem->UpdateState(switchDisplay, false);
+				oldChannelPatchItem->UpdateDisplays(mainDisplay, switchDisplay);
+			}
+			sActiveProgramChangePatches[ch] = curSwitchItem->mPatch;
 		}
 	}
 }